+++ src/intersection/intersection.rs
