+++ src/intersection/mod.rs
@@ -1,5 +1,5 @@
 pub mod compound_curve_intersection;
-pub mod curve_intersection;
+pub mod intersection;
 pub mod curve_intersection_bfgs;
 pub mod curve_intersection_problem;
 pub mod curve_intersection_solver_options;
@@ -10,7 +10,7 @@ pub mod intersection_surface_curve;
 pub mod surface_curve_intersection;
 
 pub use compound_curve_intersection::*;
-pub use curve_intersection::*;
+pub use intersection::*;
 pub use curve_intersection_bfgs::*;
 pub use curve_intersection_problem::*;
 pub use curve_intersection_solver_options::*;
